EASY-2404: Uitbreiding van easy-update-metadata-dataset: samengestelde velden + waardes toevoegen (#24)
The new element can be a simple element like 

     <dct:license>http://creativecommons.org/licenses/by/4.0</dct:license>

or it can be a complex element like 

  ```
<emd:rights>
    <dct:accessRights eas:schemeId="common.dcterms.accessrights">OPEN_ACCESS</dct:accessRights>
    <dct:license>accept</dct:license>
    <dct:license>http://creativecommons.org/licenses/by/4.0</dct:license>
  </emd:rights>
```

The new element is given as an inline parameter, and it is added to the given parent element. In this example \<dct:license\> element is added to \<emd:rights\> parent element:

  `easy-dataset:7,EMD,rights,EMPTY, 
         <dct:license>http://creativecommons.org/licenses/by/4.0</dct:license>`

This command is `idempotent` likewise the other commands in easy-update-metadata.  But you have to take the following in account:  If we would add the complex element here above (<emd:rights>), and then run the same command again, the new element would be added again.  It is because even though we don't give any attributes for <emd:rights> node, they will be deduced.  We have to give the attributes explicitly.  In this case:

```
<emd:rights xmlns:dc="http://purl.org/dc/elements/1.1/" xmlns:dct="http://purl.org/dc/terms/" xmlns:eas="http://easy.dans.knaw.nl/easy/easymetadata/eas/" xmlns:emd="http://easy.dans.knaw.nl/easy/easymetadata/" emd:version="0.1"> 
   <dct:accessRights eas:schemeId="common.dcterms.accessrights">OPEN_ACCESS</dct:accessRights>
   <dct:license>accept</dct:license>
   <dct:license>http://creativecommons.org/licenses/by/4.0</dct:license>
</emd:rights> 
```